+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Next
 
+- fix: do not try to use the view if it is in a dirty state ([#278](https://github.com/PostHog/posthog-android/pull/278))
+
 ## 3.20.3 - 2025-08-22
 
 - fix: do not try to create a Canvas with a recycled bitmap ([#277](https://github.com/PostHog/posthog-android/pull/277))